pyproject.toml
[project]
name = "vulcan-welding-agent"
version = "0.1.0"
description = "Multimodal AI assistant for the Vulcan OmniPro 220 welding machine"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.0.0",
    "pillow>=11.0.0",
    "numpy>=2.0.0",
    "scikit-learn>=1.5.0",
    "python-multipart>=0.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
